1.거래내역에 거래소 id 추가 2.거래내역 추가 기능 구현 3. 잔고테이블 반영 구현

+++ routers/yield.js
@@ -7,7 +7,8 @@ module.exports = function(){
 
     if(req.user)
     {
-      var sql = `SELECT deal_dt
+      var sql = `SELECT exchng_id
+                      , DATE_FORMAT(deal_dt, '%Y%m%d') deal_dt
                       , deal_no
                       , deal_tp
                       , crnc_code
@@ -19,7 +20,7 @@ module.exports = function(){
                    FROM trade_info
                   WHERE user_id = ?`;
 
-      conn.query(sql, [req.user.user_id], function(err, rows, fields){
+      conn.query(sql, [req.user.id], function(err, rows, fields){
 
         var db_trade_info = rows;
 
@@ -78,44 +79,131 @@ module.exports = function(){
 
   route.post('/add', function(req, res){
 
-    console.log('add', req.body);
+    if(req.user)
+    {
+      console.log('add', req.body);
+      console.log('add', req.user);
 
-    var in_deal_dt        = req.body.deal_dt;
-    var in_deal_no
-    var in_user_id        = req.user.user_id;
-    var in_deal_tp        = req.body.deal_tp;
-    var in_crnc_code      = req.body.crnc_code;
-    var in_key_crnc_code  = req.body.key_crnc_code;
-    var in_qry            = req.body.qry;
-    var in_unit_prc       = req.body.unit_prc;
-    var in_deal_amt       = req.body.deal_amt;
-    var in_cmsn_amt       = req.body.cmsn_amt;
+      var in_deal_dt        = req.body.deal_dt;
+      var in_deal_no        = 0;
+      var in_id             = req.user.id;
+      var in_exchng_id      = req.body.exchng_id;
+      var in_deal_tp        = req.body.deal_tp;
+      var in_crnc_code      = req.body.crnc_code;
+      var in_key_crnc_code  = req.body.key_crnc_code;
+      var in_qry            = req.body.qry;
+      var in_unit_prc       = req.body.unit_prc;
+      var in_deal_amt       = req.body.deal_amt;
+      var in_cmsn_amt       = req.body.cmsn_amt;
 
-    var params = [in_deal_dt, in_user_id, in_deal_tp, in_crnc_code, in_key_crnc_code, in_qry, in_unit_prc, in_deal_amt, in_cmsn_amt];
 
-    var sql = `INSERT INTO trade_info(
-                deal_dt,
-                user_id,
-                deal_tp,
-                crnc_code,
-                key_crnc_code,
-                qry,
-                unit_prc,
-                deal_amt,
-                cmsn_amt)
-               VALUES(?, ?, ?, ?, ?, ?, ? ,? ,?)`;
+      var sql = `SELECT COALESCE(MAX(deal_no), 0) + 1 deal_no
+                   FROM trade_info
+                  WHERE deal_dt = ?
+                    AND user_id = ?`;
 
-    conn.query(sql, params, function(err, rows, fields){
+      conn.query(sql, [req.body.deal_dt, req.user.id], function(err, rows, fields){
 
-      if(err){
-        console.log(err);
-        res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
-      }
-      else{
-        console.log('redirect root');
-         res.redirect('/')
-      }
-    });
+        if(err){
+          console.log(err);
+          res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+        }
+        else{
+
+          var in_deal_no = rows[0].deal_no;
+          console.log('거래번호 채번 성공 : ', in_deal_no);
+
+          var params = [in_deal_dt, in_deal_no, in_id, in_exchng_id ,in_deal_tp, in_crnc_code, in_key_crnc_code, in_qry, in_unit_prc, in_deal_amt, in_cmsn_amt];
+
+          var sql = `INSERT INTO trade_info(
+                      deal_dt,
+                      deal_no,
+                      user_id,
+                      exchng_id,
+                      deal_tp,
+                      crnc_code,
+                      key_crnc_code,
+                      qry,
+                      unit_prc,
+                      deal_amt,
+                      cmsn_amt)
+                     VALUES(?, ?, ?, ?, ?, ?, ?, ? ,? ,?, ?)`;
+
+          conn.query(sql, params, function(err, rows, fields){
+
+            if(err){
+              console.log(err);
+              res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+            }
+            else{
+              console.log('redirect root');
+
+              var sql = `SELECT *
+                           FROM balance
+                          WHERE exchng_id = ?
+                            AND crnc_code = ?
+                            AND user_id   = ?`;
+
+              conn.query(sql, [in_exchng_id, in_crnc_code, in_id], function(err, rows, fields){
+
+                if(err){
+                  console.log(err);
+                  res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+                }
+                else{
+
+                  var balance_info = rows[0];
+
+                  if(balance_info){
+
+                    if(in_deal_tp == '2')
+                    {
+                      in_qry = (-1)*in_qry;
+                    }
+
+                    var sql = `UPDATE balance
+                                  SET balance_amt = balance_amt + ?
+                                WHERE exchng_id = ?
+                                  AND crnc_code = ?
+                                  AND user_id   = ?`;
+
+                    conn.query(sql, [in_qry, in_exchng_id, in_crnc_code, in_id], function(err, rows, fields){
+
+                      if(err){
+                        console.log(err);
+                        res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+                      }
+                    });
+
+                  }
+                  else {
+                    var sql = `INSERT INTO balance(
+                                exchng_id,
+                                crnc_code,
+                                user_id,
+                                balance_amt)
+                               VALUES(?, ?, ?, ?)`;
+
+                    conn.query(sql, [in_exchng_id, in_crnc_code, in_id, in_qry], function(err, rows, fields){
+
+                      if(err){
+                        console.log(err);
+                        res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+                      }
+                    });
+                  }
+                }
+
+               res.redirect('/')
+             });
+           }
+         });
+       }
+      });
+    }
+    else {
+      res.render('auth/login');
+    }
 
     // db.select('max(deal_no)').from('traninfo')
     //   .where(
@@ -158,42 +246,48 @@ module.exports = function(){
 
   route.post('/edit', function(req, res){
 
-    var in_deal_dt        = req.body.deal_dt;
-    var in_deal_no        = req.body.deal_no;
-    var in_user_id        = req.body.user_id;
-    var in_deal_tp        = req.body.deal_tp;
-    var in_crnc_code      = req.body.crnc_code;
-    var in_key_crnc_code  = req.body.key_crnc_code;
-    var in_qry            = req.body.qry;
-    var in_unit_prc       = req.body.unit_prc;
-    var in_deal_amt       = req.body.deal_amt;
-    var in_cmsn_amt       = req.body.cmsn_amt;
+    if(req.user)
+    {
+      var in_deal_dt        = req.body.deal_dt;
+      var in_deal_no        = req.body.deal_no;
+      var in_id             = req.user.id;
+      var in_deal_tp        = req.body.deal_tp;
+      var in_crnc_code      = req.body.crnc_code;
+      var in_key_crnc_code  = req.body.key_crnc_code;
+      var in_qry            = req.body.qry;
+      var in_unit_prc       = req.body.unit_prc;
+      var in_deal_amt       = req.body.deal_amt;
+      var in_cmsn_amt       = req.body.cmsn_amt;
 
-    var params= [in_deal_tp, in_crnc_code, in_key_crnc_code, in_qry, in_unit_prc, in_deal_amt, in_cmsn_amt, in_deal_dt, in_deal_no, in_user_id];
+      var params= [in_deal_tp, in_crnc_code, in_key_crnc_code, in_qry, in_unit_prc, in_deal_amt, in_cmsn_amt, in_deal_dt, in_deal_no, in_id];
 
-    var sql = `UPDATE trade_info SET
-               deal_tp = ?,
-               crnc_code = ?,
-               key_crnc_code = ?,
-               qry = ?,
-               unit_prc = ?,
-               deal_amt = ?,
-               cmsn_amt = ?
-               WHERE deal_dt = ?
-                 AND deal_no = ?
-                 AND user_id = ?
-              `;
+      var sql = `UPDATE trade_info SET
+                 deal_tp = ?,
+                 crnc_code = ?,
+                 key_crnc_code = ?,
+                 qry = ?,
+                 unit_prc = ?,
+                 deal_amt = ?,
+                 cmsn_amt = ?
+                 WHERE deal_dt = ?
+                   AND deal_no = ?
+                   AND user_id = ?
+                `;
 
-    conn.query(sql, params, function(err, rows, fields){
+      conn.query(sql, params, function(err, rows, fields){
 
-      if(err){
-        res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
-      }
-      else{
-        res.redirect('/yield');
-      }
+        if(err){
+          res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+        }
+        else{
+          res.redirect('/yield');
+        }
 
-    })
+      })
+    }
+    else {
+      res.render('auth/login');
+    }
 
     // db.update('traninfo').set(
     //   {
@@ -239,26 +333,31 @@ module.exports = function(){
 
   route.post('/delete', function(req, res){
 
-    var in_deal_dt        = req.body.deal_dt;
-    var in_deal_no        = req.body.deal_no;
-    var in_user_id        = req.body.user_id;
+    if(req.user)
+    {
+      var in_deal_dt        = req.body.deal_dt;
+      var in_deal_no        = req.body.deal_no;
+      var in_id             = req.user.id;
 
-    var sql = `DELETE FROM trade_info
-                WHERE deal_dt = ?
-                  AND deal_no = ?
-                  AND user_id = ?
-              `;
+      var sql = `DELETE FROM trade_info
+                  WHERE deal_dt = ?
+                    AND deal_no = ?
+                    AND user_id = ?
+                `;
 
-    conn.query(sql, [in_deal_dt, in_deal_no, in_user_id], function(err, rows, fields){
+      conn.query(sql, [in_deal_dt, in_deal_no, in_id], function(err, rows, fields){
 
-      if(err){
-        res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
-      }
-      else{
-        res.redirect('/yield');
-      }
-    });
-
+        if(err){
+          res.status(500).send('문제가 발생했습니다. 빠른 시일안에 해결하겠습니다.');
+        }
+        else{
+          res.redirect('/yield');
+        }
+      });
+    }
+    else {
+          res.render('auth/login');
+    }
     // var in_id = req.params.id;
     // db.delete().from('traninfo').where(
     //   {
